services/forecaster/requirements.txt
fastapi
uvicorn
pydantic
pandas
# Use the official library from Amazon Science
chronos-forecasting